tool_pdf2markdown/requirements.txt
pdfplumber>=0.10.3
PyMuPDF>=1.23.7
Pillow>=10.0.0 